Fix organization of time-series graph classes
