+++ docs/references.bib
@@ -36,6 +36,13 @@
   doi     = {10.1007/s10618-010-0178-6}
 }
 
+@article{gerhardus2021characterization,
+  title   = {Characterization of causal ancestral graphs for time series with latent confounders},
+  author  = {Gerhardus, Andreas},
+  journal = {arXiv preprint arXiv:2112.08417},
+  year    = {2021}
+}
+
 
 @inproceedings{Malinsky18a_svarfci,
   title     = {Causal Structure Learning from Multivariate Time Series in Settings with Unmeasured Confounding},
